+++ Android/app/src/main/java/com/ssafy/kkaddak/domain/entity/market/HistoryItem.kt
@@ -1,8 +1,6 @@
 package com.ssafy.kkaddak.domain.entity.market
 
-import java.math.BigInteger
-
 data class HistoryItem(
-    val price: BigInteger,
+    val price: Double,
     val timestamp: String
 )
